utils/data_janitor/requirements.txt
pandas>=2.0.0
boto3>=1.34.0
python-dotenv>=1.0.0
pyarrow>=14.0.0  # Required for Parquet support 